% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_survival_simu_data.R
\name{generate_survival_simu_data}
\alias{generate_survival_simu_data}
\title{Generate multi-block survival simulated data.}
\usage{
generate_survival_simu_data(n, B = 2, B.CGH = 1, rho = rep(0.7, B - 1),
  p = rep(100, B), n.expl = rep(100, B), noise_corr = rep(0.5, B),
  alpha = 0.1, seed = 4257)
}
\arguments{
\item{n}{total number of individuals.}

\item{B}{number of non-CGH blocks.}

\item{B.CGH}{number of CGH blocks.}

\item{p}{vector of number of variables per block.}

\item{n.expl}{vector of number of informative variables per block.}

\item{noise_corr}{vector of correlation noise parameters per block.}

\item{alpha}{???}

\item{seed}{random seed.}
}
\value{
Simulated data: X, y, names of the variables and hazard function.
}
\description{
Generate multi-block survival simulated data.
}
\examples{
nf <- 10 # nb of folds for CV
n <- nf*nf*2 #observations
p <- c(40, 30) # c(40, 30, 50, 20, 30, 45) #nb of variables in each block
noise_corr <- c(0.8, 0.5) # correlation between variables within each block
B <- length(p) # number of blocks
rho <- 0.7 # correlation parameter !!! choose carefully the rho's so that sigmaComponents is positive definite !!!
alpha <- 0.1 or sqrt(2) # coefficient that moderates eta's part in the model
B.CGH <- 1 # nb of CGH simulated blocks
nb.expl <- 100 # nb of variables with information in each block
library(multiblog) # for idn function
}

