% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/istacox.predict.R
\name{istacox.predict}
\alias{istacox.predict}
\title{Predicts (sparse log-likelihood, prognostic index or deviance) from a model and test data.}
\usage{
istacox.predict(model, x, y, D = NULL, lambda, type = c("spll", "deviance",
  "pi"))
}
\arguments{
\item{model}{list of B beta estimates from the train set,}

\item{x}{list of B matrices of covariates of the test set,}

\item{y}{data frame of survival outcome with time and status,}

\item{D}{B by B matrix indicating the links between blocks}

\item{spll}{= sparse partial loglikelihood}

\item{deviance}{= related to the null model}

\item{pi}{= prognostic index (cf Bovelstad 2007)}
}
\value{
the quality of the prediction specified by the user.
}
\description{
Predicts (sparse log-likelihood, prognostic index or deviance) from a model and test data.
}

