% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/istacox.score.R
\name{istacox.score}
\alias{istacox.score}
\title{Performance score}
\usage{
istacox.score(y.test, y.hat)
}
\arguments{
\item{y.test}{}

\item{y.hat}{}
}
\value{
performance score
}
\description{
Performance score
}
\keyword{internal}

