% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/istacox_step_line_search.R
\name{istacox_step_line_search}
\alias{istacox_step_line_search}
\title{Computes the step for an iteration of ISTA Cox regression with a line search algorithm.}
\usage{
istacox_step_line_search(X, beta, I, R, t = 10, tau = 0.5, alpha, link,
  kmax = 1000)
}
\arguments{
\item{X}{current data block.}

\item{beta}{current beta.}

\item{I}{list of non censored individuals, ranked by event time.}

\item{R}{list of sets of individuals at risk at each non censored ranked time.}

\item{t}{is the maximum step for the line search.}

\item{tau}{is the factor for the line search.}

\item{alpha}{L1-norm shrinkage parameter.}

\item{link}{is the total link between the current block and the other blocks.}

\item{kmax}{is the maximal number of iterations.}
}
\value{
step
}
\description{
Computes the step for an iteration of ISTA Cox regression with a line search algorithm.
}

