% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Compute the link between blocks.}
\usage{
link(X, D, b, beta.init)
}
\arguments{
\item{X}{a list of B matrices}

\item{D}{is a B by B binary design matrix describing connections between blocks}

\item{b}{is the block to be treated by istacox}

\item{beta}{list of B cox coefficient vectors}
}
\value{
The sum of covariance between linked blocks
}
\description{
Compute the link between blocks.
}

