\name{MULTIBLOX-package}
\alias{MULTIBLOX-package}
\alias{MULTIBLOX}
\docType{package}
\title{
\packageTitle{MULTIBLOX}
}
\description{
\packageDescription{MULTIBLOX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MULTIBLOX}
\packageIndices{MULTIBLOX}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MULTIBLOX}

Maintainer: \packageMaintainer{MULTIBLOX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
