% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.l1.R
\name{norm.l1}
\alias{norm.l1}
\title{Computes the L1-norm of a vector}
\usage{
norm.l1(beta)
}
\arguments{
\item{beta}{is a vector of values}
}
\value{
the L1-norm of beta = sum(|beta|)
}
\description{
Computes the L1-norm of a vector
}
\keyword{internal}

