% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grad.R
\name{grad}
\alias{grad}
\title{Gradient}
\usage{
grad(X, beta, I, R, alpha, link)
}
\arguments{
\item{X}{current data block.}

\item{beta}{current beta.}

\item{I}{list of non censored individuals, ranked by event time.}

\item{R}{list of sets of individuals at risk at each non censored ranked time.}

\item{alpha}{L1-norm shrinkage parameter.}

\item{link}{is the total link between the current block and the other blocks.}
}
\value{
gradient of Cox partial likelihood
}
\description{
Gradient
}
\keyword{internal}

