% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istacox.R
\name{istacox}
\alias{istacox}
\title{Computes the beta coefficients for a sparse cox model via ISTA}
\usage{
istacox(X, I, R, alpha, kmax = 1000, epsilon = 1e-04, fast = FALSE,
  ada = FALSE, link, beta_init)
}
\arguments{
\item{X}{is a list of B matrices (blocks)}

\item{I}{list of non censored individuals, ranked by event time.}

\item{R}{list of sets of individuals at risk at each non censored ranked time}

\item{alpha}{L1-norm shrinkage parameter}

\item{kmax}{is the maximal number of iterations.}

\item{epsilon}{is the required precision.}

\item{fast}{is a boolean used to specify if FISTA (TRUE) is used instead of ISTA (FALSE, default).}

\item{ada}{is a boolean used to specify is the step must be chosen at each iteration (TRUE) or not (FALSE, default).}

\item{link}{is the total link between the current block and the other blocks.}

\item{beta_init}{is the initial value of the weight vector (used mainly for warm restarts).}
}
\value{
beta coefficients that maximize the Cox sparse partial likekihood, and k number of iterations
}
\description{
Computes the beta coefficients for a sparse cox model via ISTA
}

