% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm.l1.R
\name{norm.l1}
\alias{norm.l1}
\title{Computes the l1 norm of a vector.}
\usage{
norm.l1(beta)
}
\arguments{
\item{beta}{vector.}
}
\value{
sum |beta_i|
}
\description{
Computes the l1 norm of a vector.
}
\keyword{internal}

